学校学院

+++ src/main/java/com/bootdo/thesisMamager/controller/ThesisCollegeController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import java.util.Map;
 
 import com.bootdo.common.utils.*;
+import com.bootdo.thesisMamager.service.ThesisStudentService;
+import com.bootdo.thesisMamager.service.ThesisTeacherService;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
@@ -21,25 +23,30 @@ import com.bootdo.thesisMamager.domain.ThesisCollegeDO;
 import com.bootdo.thesisMamager.service.ThesisCollegeService;
 
 /**
- * 
- * 
+ *
+ *
  * @author xc
  * @email 1992lcg@163.com
  * @date 2018-01-04 16:03:33
  */
- 
+
 @Controller
 @RequestMapping("/thesisMamager/thesisCollege")
 public class ThesisCollegeController {
 	@Autowired
 	private ThesisCollegeService thesisCollegeService;
-	
+	@Autowired
+	private ThesisTeacherService teacherService;
+	@Autowired
+	private ThesisStudentService thesisStudentService;
+
+
 	@GetMapping()
 	@RequiresPermissions("thesisMamager:thesisCollege:thesisCollege")
 	String ThesisCollege(){
 	    return "thesisMamager/thesisCollege/thesisCollege";
 	}
-	
+
 	@ResponseBody
 	@GetMapping("/list")
 	@RequiresPermissions("thesisMamager:thesisCollege:thesisCollege")
@@ -51,7 +58,7 @@ public class ThesisCollegeController {
 
 		return thesisCollegeList;
 	}
-	
+
 	@GetMapping("/add/{pId}")
 	@RequiresPermissions("thesisMamager:thesisCollege:add")
 	String add(@PathVariable("pId") String pId, Model model) {
@@ -79,7 +86,7 @@ public class ThesisCollegeController {
 		model.addAttribute("thesisCollege", thesisCollege);
 	    return "thesisMamager/thesisCollege/edit";
 	}
-	
+
 	/**
 	 * 保存
 	 */
@@ -113,20 +120,51 @@ public class ThesisCollegeController {
 		thesisCollegeService.update(thesisCollege);
 		return R.ok();
 	}
-	
+
 	/**
 	 * 删除
 	 */
 	@PostMapping( "/remove")
 	@ResponseBody
 	@RequiresPermissions("thesisMamager:thesisCollege:remove")
-	public R remove( Long id){
-		if(thesisCollegeService.remove(id)>0){
-		return R.ok();
+	public R remove( Long id) {
+		Map map = new HashMap();
+		map.put("schoolId", id);
+		Map map2 = new HashMap();
+		map2.put("depId", id);
+		if (thesisCollegeService.getByPid(id) != null) {
+			return R.error("删除失败,下级还有院系数据");
+		}
+		if (teacherService.list(map).size() > 0 || thesisStudentService.list(map).size() > 0) {
+			return R.error("删除失败,下级还有导师和学生的数据");
+		}
+		if (teacherService.list(map2).size() > 0 || thesisStudentService.list(map2).size() > 0) {
+			return R.error("删除失败,下级还有导师和学生的数据");
+		}
+		if (thesisCollegeService.remove(id) > 0) {
+			return R.ok();
 		}
 		return R.error();
 	}
-	
+
+	/**
+	 * stop
+	 */
+	@PostMapping( "/stop")
+	@ResponseBody
+	@RequiresPermissions("thesisMamager:thesisCollege:stop")
+	public R stop( Long id) {
+		ThesisCollegeDO thesisCollegeDO = new ThesisCollegeDO();
+		thesisCollegeDO.setId(id);
+		if(thesisCollegeService.get(id).getState() == 0){
+			thesisCollegeDO.setState(1);
+		}else{
+			thesisCollegeDO.setState(0);
+		}
+		thesisCollegeService.update(thesisCollegeDO);
+		return R.ok();
+	}
+
 	/**
 	 * 删除
 	 */
@@ -137,5 +175,5 @@ public class ThesisCollegeController {
 		thesisCollegeService.batchRemove(ids);
 		return R.ok();
 	}*/
-	
+
 }
